Prevent terminated `types.SessionTracker` from lingering
Transitioning to terminated didn't have any impact on the TTL of
the resource or cause it to be deleted. This resulted in having
terminated `types.SessionTracker` resources exist up to an hour
after they had been terminated.

To remedy this and prune terminated `types.SessionTrackers` more
frequently, the TTL is reduced to 3m upon changing state to terminated.

Fixes #16644