Support tctl auth sign --format=windows
User certs for desktop access are only valid for a few minutes
and are never written to disk. This can make it difficult to
troubleshoot cert validity.

This commit adds support for generating Windows user certificates
which can be exported to the Windows environment for validation.

Note: at this time, we only write the certificate and not the
corresponding private key, making it impossible to use the generated
cert for any real purpose.